Fade effect

+++ src/gameplay.js
@@ -8,6 +8,7 @@ var Gameplay = function () {
   this.mapKey = null;
   this.directionFrom = null;
   this.nodeMap = null;
+  this.fading = false;
 };
 Gameplay.prototype.init = function (mapKey, directionFrom) {
   this.mapKey = mapKey;
@@ -39,6 +40,7 @@ Gameplay.prototype.create = function() {
   this.player = this.game.add.existing(new Player(this.game, entryPoint ? entryPoint.x : 128, entryPoint ? entryPoint.y : 128));
   this.player.renderable = false;
   this.player.targetPt.renderable = false;
+  this.player.facingDirection = this.directionFromToPlayerDirection(this.directionFrom);
 
   this.player.events.onKilled.add(function () {
     this.game.time.events.add(2000, function () {
@@ -56,17 +58,21 @@ Gameplay.prototype.create = function() {
     wolf.renderable = false;
   }, this);
   
-  this.game.add.bitmapText(32, 32, 'font', 'scene 3', 8);
+  this.game.add.bitmapText(32, 32, 'font', this.mapKey, 8);
 
   setupThreeScene(this.game, this.player, this.wolves);
+
+  this.game.camera.flash(0x3a4d51, 700);
 };
 Gameplay.prototype.update = function() {
   this.game.physics.arcade.collide(this.player, this.foreground);
   this.game.physics.arcade.collide(this.wolves, this.foreground);
 
-  this.game.physics.arcade.overlap(this.player, this.wolves, function (player, wolf) {
-    player.kill();
-  }, undefined, this);
+  if (this.fading === false) {
+    this.game.physics.arcade.overlap(this.player, this.wolves, function (player, wolf) {
+      player.kill();
+    }, undefined, this);
+  }
 
   if (this.player.currentState === PlayerState.ATTACK) {
     this.game.physics.arcade.overlap(this.player.targetPt, this.wolves, function (targetPt, wolf) {
@@ -84,18 +90,56 @@ Gameplay.prototype.update = function() {
   UpdateThreeScene(this.player, this. wolves);
 
   // transition to other maps
-  if (this.player.y < 32 && this.map.properties.northExit !== undefined) {
-    this.game.state.start('Gameplay', true, false, this.map.properties.northExit, 'north');
-  } else if (this.player.y > (this.map.heightInPixels - 32) && this.map.properties.southExit !== undefined) {
-    this.game.state.start('Gameplay', true, false, this.map.properties.southExit, 'south');
-  } else if (this.player.x < 32 && this.map.properties.westExit !== undefined) {
-    this.game.state.start('Gameplay', true, false, this.map.properties.westExit, 'west');
-  } else if (this.player.x > (this.map.widthInPixels - 32) && this.map.properties.eastExit !== undefined) {
-    this.game.state.start('Gameplay', true, false, this.map.properties.eastExit, 'east');
+  if (this.fading === false) {
+    var exit = undefined;
+    var exitDirection = undefined;
+
+    if (this.player.y < 32 && this.map.properties.northExit !== undefined) {
+      exit = this.map.properties.northExit
+      exitDirection = 'north';
+    } else if (this.player.y > (this.map.heightInPixels - 32) && this.map.properties.southExit !== undefined) {
+      exit = this.map.properties.southExit
+      exitDirection = 'south';
+    } else if (this.player.x < 32 && this.map.properties.westExit !== undefined) {
+      exit = this.map.properties.westExit
+      exitDirection = 'west';
+    } else if (this.player.x > (this.map.widthInPixels - 32) && this.map.properties.eastExit !== undefined) {
+      exit = this.map.properties.eastExit
+      exitDirection = 'east';
+    }
+
+    if (exit !== undefined && exitDirection !== undefined) {
+      this.game.camera.onFadeComplete.add(function () {
+        this.fading = false;
+        this.game.state.start('Gameplay', true, false, exit, exitDirection);
+        this.game.camera.onFadeComplete.removeAll();
+      }, this);
+      this.fading = true;
+      this.game.camera.fade(0x3a4d51, 1000, true);
+      this.player.update = function() {}; // quick/dirty hack to prevent player movement
+    }
   }
 };
 Gameplay.prototype.preRender = function () {
   ThreeRenderer.render(ThreeScene, ThreeCamera);
+};
+Gameplay.prototype.directionFromToPlayerDirection  = function(directionFrom) {
+  switch (directionFrom) {
+    case 'east':
+      return new Phaser.Point(1, 0);
+    break;
+    case 'south':
+      return new Phaser.Point(0, 1);
+    break;
+    case 'west':
+      return new Phaser.Point(-1, 0);
+    break;
+    case 'north':
+      return new Phaser.Point(0, -1);
+    break;
+  }
+
+  return new Phaser.Point(0, 1);
 };
 Gameplay.prototype.shutdown = function() {
   UnloadThreeScene(this.wolves);
@@ -109,4 +153,5 @@ Gameplay.prototype.shutdown = function() {
   this.mapKey = null;
   this.directionFrom = null;
   this.nodeMap = null;
+  this.fading = false;
 };
